TIP-841: Move the channel translation model to Akeneo\Channel
